+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 current - 2018-04-15 - 0.00Mb
 =============================
 
+* `3`: add ip address in the logging (2018-04-15)
 * `2`: add encrypted logging (2018-04-15)
 * `1`: fix gallery of examples (style) (2018-01-05)